Renamed PetsFilter to PetFilter
